zero: Use the same module for both vfs versions
